+++ examples/Cy-Pt@graphene/autoneb/autoneb_run.py
@@ -2,10 +2,8 @@
 # Run AutoNEB calculation on NEB images by ASE-ABACUS
 # part of ASE-NEB-ABACUS scripts
 
-import os
-from ase.calculators.abacus import Abacus, AbacusProfile
 from ase.optimize import FIRE, BFGS
-from ase.mep.autoneb import AutoNEB
+from abacus_autoneb import AbacusAutoNEB
 from ase.io import read, write
 from ase.parallel import world, parprint, paropen
 #from pathlib import Path
@@ -70,89 +68,6 @@ parameters = {
     'efield_dir': 1,
     'efield_pos_max': 0.0
 }
-
-
-class AbacusAutoNEB:
-    """Customize AutoNEB calculation by using ABACUS"""
-
-    def __init__(self, init_chain, parameters, abacus='abacus',  
-                 prefix="run_autoneb", n_simul=1, n_max=10, algorism="improvedtangent", 
-                 directory='NEB', mpi=1, omp=1, parallel=True, ):
-        """Initialize initial and final states
-
-        init_chain (Atoms object): starting image chain from nem_make.py or other method, can only include Initial and Final states, autoneb will generate inter-images automactically. 
-        parameters (dict): settings of abacus input parameters
-        abacus (str): Abacus executable file. Default: 'abacus'
-        algorism (str): NEB algorism. which can be
-        - 'aseneb': standard ase NEB
-        - 'improvedtangent' : IT-NEB (recommended by Sobereva)
-        - 'eb': climbing image elastic band method (default in AutoNEB)
-        
-        Default: 'improvedtangent'
-        
-        prefix (str): prefix for AutoNEB output files, default 'autoneb'
-        n_simul (int): number of simultaneous calculation, default use world.size, read from ase.parallel
-        n_max (int): max number of all image in NEB band, default 10. The max number will not be reached if the convergence is reached before that.
-        directory (str): calculator directory name, for parallel calculation {directory}-rank{i} will be the directory name
-        mpi (int): number of MPI for abacus calculator
-        omp (int): number of OpenMP for abacus calculator
-        parallel (bool): parallel calculation setting, default True
-        """
-
-        self.init_chain = init_chain
-        self.algorism = algorism
-        self.abacus = abacus
-        self.directory = directory
-        self.parameters = parameters
-        self.prefix = prefix
-        self.mpi = mpi
-        self.omp = omp
-        self.parallel = parallel
-        parprint("Notice: AutoNEB method is set")
-        if (n_simul > 0) and (n_max >= n_simul):
-            parprint(f"You manually set n_simul = {n_simul}, n_max = {n_max}", )
-            self.n_simul = n_simul
-            self.n_max = n_max
-        else:
-            raise ValueError("You must set n_simul > 0 and n_max >= n_simul numbers for AutoNEB")
-        
-            
-    def set_calculator(self):
-        """Set Abacus calculators"""
-        os.environ['OMP_NUM_THREADS'] = f'{self.omp}'
-        profile = AbacusProfile(
-            argv=['mpirun', '-np', f'{self.mpi}', self.abacus])
-        if self.parallel:
-            out_directory = f"{self.directory}-rank{world.rank}"
-        else:
-            out_directory = self.directory
-        calc = Abacus(profile=profile, directory=out_directory,
-                    **self.parameters)
-        return calc
-    
-    
-    def attach_calculators(self, images: list):
-        """Attach calculator to a list of images supplied"""
-        for num, image in enumerate(images):
-            image.calc = self.set_calculator()
-        
-    
-    def run(self, optimizer=FIRE, fmax=0.05, climb=True):
-        """Run Abacus AutoNEB
-
-        optimizer (Optimizer object): defaults to FIRE. BFGS and FIRE is only used
-        fmax (float): threshold (unit: eV/Angstrom) of the force convergence
-        climb (bool): climbing image NEB method
-        """
-        parprint("----- Running AutoNEB -----")
-        parprint(f"----- {self.algorism} method is being used -----")
-        for num, image in enumerate(self.init_chain):
-            index = f"{num:03d}"
-            write(f"{self.prefix}{index}.traj", image, format="traj")
-        autoneb = AutoNEB(self.attach_calculators, self.prefix, self.n_simul, self.n_max, fmax=fmax, climb=climb, 
-                            method=self.algorism, parallel=self.parallel, optimizer=optimizer, )
-        autoneb.run()
-        parprint("----- AutoNEB calculation finished -----")
 
 
 if __name__ == "__main__":
